{"types":"C","runs":13}

+++ problem15.java
@@ -7,7 +7,7 @@ public class problem15 extends ConsoleProgram {
 		int numberOne = readInt("Enter First Number: ");
 		int numberSecond = readInt("Enter Second Number: ");
 		double x = 0.5;
-		double total = ((numberOne + numberSecond)/2 );
+		int total = ((numberOne + numberSecond)/2 );
 		double y = total + 0.5;
 		if (numberOne%2==0 && numberSecond%2==0) {
 		println("The answer is ->" + total + ".");
